--- src/components/HomePage/Footer/style.css
@@ -1,8 +1,0 @@
-.page-footer {
-  background-color: "grey";
-  width: "100%";
-}
-
-a {
-  color: "black";
-}
