+++ src/components/SignUpPage/index.js
@@ -1,5 +1,4 @@
 import React from "react";
-import NavBar from "./NavBar";
 import SignUpForm from "./SignUpForm";
 
 export default function SignUpPage() {
